flesh out db bits
