+++ device-aggregator/src/env.rs
@@ -1,0 +1,6 @@
+use std::env;
+
+/// Get the environment variable or panic
+pub fn get_var(name: &str) -> String {
+    env::var(name).unwrap_or_else(|_| format!("{} environment variable is required.", name))
+}
