+++ device-aggregator/src/lib.rs
@@ -2,10 +2,11 @@
 // Use of this source code is governed by a MIT-style
 // license that can be found in the LICENSE file.
 
+#[macro_use]
+extern crate diesel;
 extern crate daggy;
 extern crate derive_more;
 extern crate device_types;
-extern crate diesel;
 extern crate futures;
 extern crate im;
 extern crate libzfs_types;
@@ -16,3 +17,6 @@ extern crate tokio;
 pub mod aggregator_error;
 pub mod cache;
 pub mod dag;
+pub mod db;
+pub mod env;
+pub mod schema;
